Change other files to point to moved resources
