+++ src/utils/template/Template.ts
@@ -1,4 +1,4 @@
-import * as template from './template.html';
+import * as template from '../../../assets/template.html';
 
 declare const __LINK_LOCATION__: string;
 declare const __NEON_VERSION__: string;
